fix for RAM to be processed with double instead of ulong

+++ Dtos/RamUsageDto.cs
@@ -1,6 +1,6 @@
 namespace WorkspaceMonitor.Dtos;
 
 public record RamUsageDto(
-    ulong PercentUsage,
+    double PercentUsage,
     long Timestamp
     );
